requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.5
django-tastypie==0.14.3
pycodestyle==2.5.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
